requirements.txt
# Core API clients
openai>=1.0.0
anthropic>=0.7.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=3.1.0

# Statistical analysis
scipy>=1.10.0
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and validation
pydantic>=2.0.0
PyYAML>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0